Implement initial database system using IndexedDB
Take note that using IndexedDB is not recommended for an account authentication system because of the security risks it poses. We chose IndexedDB as our client-side database because this website is dedicated for local or offline use, as well as for the sake of implementing an easy sign up and login system. From a security standpoint, this is not recommended. Implementing this in your own projects can pose security risks.